pyproject.toml
[tool.poetry]
name = "telegram-job-bot"
version = "0.1.0"
description = "Telegram job recommendation bot integrating Adzuna and adaptive keyword learning"
authors = ["Your Name <you@example.com>"]
packages = [{include = "telegram_job_bot", from = "src"}]
readme = "telegram_job_bot_project.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
requests = "^2.31"
SQLAlchemy = "^2.0"
pydantic = "^2.5"
python-dotenv = "^1.0"
tzdata = "^2024.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
